+++ assets/js/components/index.js
@@ -1,6 +1,3 @@
-// common (reusable) components are listed here
-
-export { default as Flash } from "./flash"
 export { default as Header } from "./header"
 export { default as Loading } from "./loading"
 export { default as PRoute } from "./proute"