+++ assets/js/modules/flash/actions.js
@@ -1,10 +1,9 @@
 import types from "./types"
 
 const clear = () => ({ type: types.CLEAR })
+const error = message => ({ type: types.ERROR, message })
+const info = message => ({ type: types.INFO, message })
+const success = message => ({ type: types.SUCCESS, message })
+const warning = message => ({ type: types.WARNING, message })
 
-const error = ( content, header="An Error Occurred!" ) => ({ type: types.ERROR, header, content })
-const info = ( content, header="Please note:" ) => ({ type: types.INFO, header, content })
-const success = (content, header="Success!" ) => ({ type: types.SUCCESS, header, content })
-const warning = ( content, header="Warning:" ) => ({ type: types.WARNING, header, content })
-
-export default { clear, error, info, success, warning };
+export default { clear, error, info, success, warning }