+++ assets/js/modules/flash/index.js
@@ -1,4 +1,6 @@
-import reducer from "./reducers"
+import reducer from './reducers'
 
-export { default as flashOperations } from "./operations"
-export default reducer;
+export { default as flashOperations } from './operations'
+export { default as flashSelectors } from './selectors'
+
+export default reducer